Move tests out of bitcoindRpcTest that belong in async-utils
